Removed unused shares from Dataset type and move queries from dataset and dataset_table into dataset_sharing
