+++ backend/dataall/modules/dataset_sharing/api/queries.py
@@ -4,6 +4,8 @@ from dataall.modules.dataset_sharing.api.resolvers import (
     list_shared_with_environment_data_items,
     list_shares_in_my_inbox,
     list_shares_in_my_outbox,
+    list_dataset_share_objects,
+    list_shared_tables_by_env_dataset,
 )
 
 getShareObject = gql.QueryField(
@@ -37,4 +39,25 @@ searchEnvironmentDataItems = gql.QueryField(
     resolver=list_shared_with_environment_data_items,
     type=gql.Ref('EnvironmentPublishedItemSearchResults'),
     test_scope='Dataset',
+)
+
+listShareObjects = gql.QueryField(
+    name='listDatasetShareObjects',
+    resolver=list_dataset_share_objects,
+    args=[
+        gql.Argument(name='datasetUri', type=gql.NonNullableType(gql.String)),
+        gql.Argument(name='environmentUri', type=gql.String),
+        gql.Argument(name='page', type=gql.Integer),
+    ],
+    type=gql.Ref('ShareSearchResult'),
+)
+
+getSharedDatasetTables = gql.QueryField(
+    name='getSharedDatasetTables',
+    args=[
+        gql.Argument(name='datasetUri', type=gql.NonNullableType(gql.String)),
+        gql.Argument(name='envUri', type=gql.NonNullableType(gql.String)),
+    ],
+    type=gql.ArrayType(gql.Ref('SharedDatasetTableItem')),
+    resolver=list_shared_tables_by_env_dataset,
 )
