+++ frontend/src/services/graphql/Datasets/index.js
@@ -2,5 +2,4 @@ export * from './addDatasetStorageLocation';
 export * from './getDataset';
 export * from './getDatasetAssumeRoleUrl';
 export * from './listDatasetTables';
-export * from './listShareObjects';
 export * from './removeDatasetStorageLocation';
